requirements.txt
appdirs==1.4.4
astroid==2.4.2
attrs==20.1.0
aws-cdk.assets==1.61.1
aws-cdk.aws-applicationautoscaling==1.61.1
aws-cdk.aws-autoscaling-common==1.61.1
aws-cdk.aws-cloudwatch==1.61.1
aws-cdk.aws-codeguruprofiler==1.61.1
aws-cdk.aws-ec2==1.61.1
aws-cdk.aws-efs==1.61.1
aws-cdk.aws-events==1.61.1
aws-cdk.aws-iam==1.61.1
aws-cdk.aws-kms==1.61.1
aws-cdk.aws-lambda==1.61.1
aws-cdk.aws-logs==1.61.1
aws-cdk.aws-s3==1.61.1
aws-cdk.aws-s3-assets==1.61.1
aws-cdk.aws-sqs==1.61.1
aws-cdk.aws-ssm==1.61.1
aws-cdk.cloud-assembly-schema==1.61.1
aws-cdk.core==1.61.1
aws-cdk.cx-api==1.61.1
aws-cdk.region-info==1.61.1
boto3==1.14.20
botocore==1.17.20
cattrs==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
constructs==3.0.4
cycler==0.10.0
decorator==4.4.2
distlib==0.3.0
docutils==0.15.2
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==1.7.0
importlib-resources==3.0.0
isort==5.4.2
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
jmespath==0.10.0
jsii==1.12.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.0
mccabe==0.6.1
moviepy==1.0.3
networkx==2.4
numpy==1.19.0
opencv-python==4.3.0.36
parso==0.7.0
pika==1.1.0
Pillow==7.2.0
pluggy==0.13.1
proglog==0.1.9
publication==0.0.3
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.33.3
PyWavelets==1.1.1
requests==2.24.0
s3transfer==0.3.3
scikit-image==0.17.2
scipy==1.5.1
six==1.15.0
tifffile==2020.7.17
toml==0.10.1
torch==1.5.1
tqdm==4.48.0
typing-extensions==3.7.4.3
ujson==1.35
urllib3==1.25.9
virtualenv==20.0.23
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
